+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 - Sortable venue roles, to determine presentation order.
 - Fixed #24, redesigning venue page to integrate roles.
 - Split volunteers page by role, in role order.
+- Improved volunteers breadcrumbs.
 
 ### Fixed